backend/requirements.txt
fastapi
uvicorn
pandas
openpyxl
plotly
numpy
pydantic-settings
python-dotenv
python-multipart
langchain
langchain-community
langchain-google-genai
matplotlib
